# openapi.yaml
openapi: 3.0.1
info:
  title: Clinical Trials Plugin
  description: A plugin that allows the user to find clinical trials using ChatGPT. If you do not know the user's username, ask them first before making queries to the plugin. Otherwise, use the username "global".
  version: 'v1'
servers:
  - url: https://openaiplugin.ngrok.app
paths:
  /clinicaltrials/{patientId}:
    post:
      operationId: getTrials
      summary: Get the list of trials
      parameters:
      - in: path
        name: patientId
        schema:
            type: string
        required: true
        description: The patient id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/trialsSearch'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTrialsResponse'
components:
  schemas:
    getTrialsResponse:
      type: object
      properties:
        trials:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                description: The id of the trial.
              description:
                type: string
                description: The description of the trial.

    trialsSearch:
      type: object
      properties:
        stage: 
          type: string
          description: The stage of the cancer.
        location:
          type: string
          description: The location of the trial
        age:
          type: string
          description: The age of the patient.
        gender:
          type: string
          description: The gender of the patient.
        targetTreatment:
          type: string
          description: The primary reason the patient is seeking treatment.
        conditions:
          type: string
          description: medical condition or medication the patient is taking.
        